Beaver .deb 64bit
Here is a .deb of the Beaver text editor for 64 bit computers. It's actually fully functioning and the plugins actually work! Unlike the version available in the Debian repository. Where the plugins are not available.  

to install on a debian/Ubuntu  distro, either download and install with gdebi-gtk or type into the terminal   'sudo dpkg -i beaver.deb '